Agrega:
- Carga logica del mapa al servidor
- Envío de mapa al cliente a traves del socket
- Carga de assets en el cliente
Arregla:
- Error de guardado/cargado de mapas en el editor
- Bug en cargado de tiles en el cliente
Modifica:
- Sprites para poder ser creados con una sola animación
- Editor para poder cargar tiles que luego serán animadas
- Parte del código para la creación de creatures en el merge resolve, para que se adapte a las modificaciones de SpriteCreator

+++ editor/view/asset_selector.cpp
@@ -6,7 +6,7 @@
 #include "ui_assetselector.h"
 
 AssetSelector::AssetSelector(const QHash<uint8_t, AssetData>& tiles,
-                             const QHash<uint16_t, AssetData>& colliders,
+                             const QHash<uint8_t, AssetData>& colliders,
                              const QHash<uint8_t, AssetData>& npcs, QWidget* parent):
         QTabWidget(parent), ui(new Ui::AssetSelector), tiles(tiles), colliders(colliders), npcs(npcs) {
